requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==24.2.0
blinker==1.9.0
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
folium==0.18.0
fonttools==4.55.1
geographiclib==2.0
gitdb==4.0.11
GitPython==3.1.43
googlemaps==4.10.0
graphviz==0.20.3
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
plotly==5.24.1
polyline==2.0.2
prompt_toolkit==3.0.48
protobuf==5.29.0
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.0
seaborn==0.13.2
setuptools==65.5.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.40.2
streamlit_folium==0.23.2
streamlit-option-menu==0.4.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
xyzservices==2024.9.0
